#62: added comments and JSDoc function signatures
