+++ lib/schema/index.js
@@ -15,9 +15,17 @@ const OUTPUT_DIR = 'reso-schema-validation-temp';
 const ERROR_REPORT = 'schema-validation-report.json';
 
 /**
- * TODO
- * @param {*} param0
- * @returns
+ * @param {Object} obj
+ * @param {boolean} obj.additionalProperties
+ * @param {boolean} obj.generate
+ * @param {boolean} obj.validate
+ * @param {string} obj.metadataPath
+ * @param {string} obj.outputPath
+ * @param {string} obj.ddVersion
+ * @param {string} obj.payloadPath
+ * @param {string} obj.resourceName
+ *
+ * @description Action handler for the commander CLI `schema` command.
  */
 const schema = async ({
   metadataPath = '',
@@ -122,7 +130,10 @@ const schema = async ({
 
 /**
  * Generates JSON schema from the given metadata
- * @param {Object} args
+ * @param {Object} obj
+ * @param {boolean} obj.additionalProperties
+ * @param {string} obj.metadataPath
+ * @param {{}} obj.metadataReport
  * @returns generated schema
  */
 const generateSchemaFromMetadata = async ({ metadataPath = '', additionalProperties = false, metadataReport }) => {
@@ -146,8 +157,15 @@ const generateSchemaFromMetadata = async ({ metadataPath = '', additionalPropert
 };
 
 /**
- * TODO
- * @param {*} param0
+ *
+ * @param {Object} obj
+ * @param {boolean} obj.additionalProperties
+ * @param {string} obj.metadataPath
+ * @param {string} obj.payloadPath
+ * @param {string} obj.version
+ * @param {string} obj.resourceName
+ *
+ * @description Reads the directory with the flattened JSON files and sends them for validation.
  */
 const validatePayloads = async ({ metadataPath = '', payloadPath = '', additionalProperties = false, version, resourceName }) => {
   const files = await readDirectory(payloadPath);
@@ -166,8 +184,15 @@ const validatePayloads = async ({ metadataPath = '', payloadPath = '', additiona
 
 /**
  *
- * @param {TODO} param0
- * @returns
+ * @param {Object} obj
+ * @param {boolean} obj.additionalProperties
+ * @param {string} obj.errorPath
+ * @param {string} obj.version
+ * @param {string} obj.metadataPath
+ * @param {string} obj.resourceName
+ * @param {string[]} obj.payloadPaths
+ *
+ * @description Processes the input from the CLI and writes the final error report into the given path.
  */
 const validatePayloadAndGenerateResults = async ({
   payloadPaths,
