+++ lib/schema/utils.js
@@ -6,9 +6,11 @@ const path = require('path');
 const { extractFilesFromZip } = require('../../common');
 
 /**
- * TODO
- * @param {*} param0
- * @returns
+ * @param {Object} obj
+ * @param {string} obj.inputPath
+ * @param {string} obj.outputPath
+ *
+ * @description Takes an input path as a JSON file, directory, or zip file. Flattens all nested JSOn files into the output directory.
  */
 const processFiles = async ({ inputPath, outputPath }) => {
   try {
@@ -35,8 +37,11 @@ const processFiles = async ({ inputPath, outputPath }) => {
 };
 
 /**
- * TODO
- * @param {*} param0
+ * @param {Object} obj
+ * @param {string} obj.filePath
+ * @param {string} obj.outputPath
+ *
+ * @description Process JSON and zip files and copy them to the output path
  */
 const processFile = async ({ filePath, outputPath }) => {
   const ext = path.extname(filePath);
@@ -54,9 +59,8 @@ const processFile = async ({ filePath, outputPath }) => {
 };
 
 /**
- * TODO - this already exists elsewhere in a common library?
- * @param {*} filePath
- * @returns
+ * @param {string} filePath
+ * @returns Contents read from the file. Returns null if file isn't present.
  */
 const readFile = async filePath => {
   try {
@@ -68,9 +72,9 @@ const readFile = async filePath => {
 };
 
 /**
- * TODO - should be moved to common library if it doesn't already exist
- * @param {*} dirName
- * @returns
+ * @param {string} dirName
+ *
+ * @description Checks if a directory exists. If not it creates the directory. If there was an error creating the directory, it returns null otherwise a truthy value.
  */
 const createDirectoryIfNotPresent = async dirName => {
   try {
@@ -86,10 +90,10 @@ const createDirectoryIfNotPresent = async dirName => {
 };
 
 /**
- * TODO - this already exists elsewhere?
- * @param {*} path
- * @param {*} data
- * @returns
+ * @param {string} path
+ * @param {string} data
+ *
+ * @description Attempts to write data intp a given file path. Returns tru on success and false in case of a failure.
  */
 const writeFile = async (path, data) => {
   try {
