+++ lib/schema/validate.js
@@ -4,6 +4,11 @@ const Ajv = require('ajv');
 const addFormats = require('ajv-formats');
 const chalk = require('chalk');
 const ajvErrors = require('ajv-errors');
+const { parseResoUrn } = require('../../common');
+
+/**
+ * @typedef {import('ajv').ValidateFunction} ValidateFunction
+ */
 
 const ajv = new Ajv({ allErrors: true, coerceTypes: false, strict: true });
 addFormats(ajv);
@@ -70,8 +75,15 @@ const isValidDdVersion = (version = '') => getValidDdVersions()?.includes(versio
 
 /**
  * TODO
- * @param {*} param0
- * @returns
+ * @param {Object} obj
+ * @param {string} obj.resourceName
+ * @param {string} obj.fileName
+ * @param {string} obj.ddVersion
+ * @param {{}} obj.jsonSchema
+ * @param {{}} obj.jsonPayload
+ * @param {{}} obj.errorMap
+ *
+ * @returns Intermediate error and warning caches along with stats. Can be combined later using `combineErrors`
  */
 const validate = ({ jsonSchema = {}, resourceName = 'Property', jsonPayload, errorMap = {}, fileName = '', ddVersion }) => {
   const { stats = { totalErrors: 0, totalWarnings: 0 }, cache = {}, errorCache = {}, warningsCache = {}, payloadErrors = {} } = errorMap;
@@ -200,8 +212,13 @@ const validate = ({ jsonSchema = {}, resourceName = 'Property', jsonPayload, err
 
 /**
  * TODO
- * @param {*} param0
- * @returns
+ * @param {Object} obj
+ * @param {Record<string, {}>} obj.payloads
+ * @param {{}} obj.schema
+ * @param {string} obj.resourceNameFromArgs
+ * @param {string} obj.versionFromArgs
+ *
+ * @returns The processed error report
  */
 const validatePayload = ({ payloads = {}, schema, resourceNameFromArgs = '', versionFromArgs }) => {
   const payloadErrors = {};
@@ -242,9 +259,14 @@ const validatePayload = ({ payloads = {}, schema, resourceNameFromArgs = '', ver
 };
 
 /**
- * TODO
- * @param {*} param0
- * @returns
+ * @param {Object} obj
+ * @param {{}} obj.cache
+ * @param {{}} obj.errorCache
+ * @param {{}} obj.warningsCache
+ * @param {{}} obj.payloadErrors
+ * @param {{totalWarnings: Number, totalErrors: Number}} obj.stats
+ *
+ * Combines results from one or more validations into one object
  */
 const combineErrors = ({ cache, errorCache, warningsCache, payloadErrors, stats }) => {
   const errorReport = [];
@@ -349,7 +371,6 @@ const combineErrors = ({ cache, errorCache, warningsCache, payloadErrors, stats 
   return {
     description: 'RESO Common Format Schema Validation Summary',
     generatedOn: new Date().toISOString(),
-    /** TODO: inferred version - this should probably just be passed in? **/
     totalErrors: stats.totalErrors,
     totalWarnings: stats.totalWarnings,
     items: errorReport
@@ -357,8 +378,17 @@ const combineErrors = ({ cache, errorCache, warningsCache, payloadErrors, stats 
 };
 
 /**
- * TODO
- * @param {*} param0
+ * @param {Object} obj
+ * @param {boolean} obj.additionalPropertiesAllowed
+ * @param {string} obj.resourceName
+ * @param {string} obj.fileName
+ * @param {{}} obj.cache
+ * @param {{}} obj.errorCache
+ * @param {{}} obj.warningsCache
+ * @param {ValidateFunction} obj.validate
+ * @param {{totalWarnings: Number, totalErrors: Number}} obj.stats
+ *
+ * Processes the raw validation results from AJV.
  */
 const generateErrorReport = ({
   validate,
@@ -394,7 +424,6 @@ const generateErrorReport = ({
       cache[resourceName][failedItemName] = {};
     }
 
-    // needs to be generic to accommodate failures that need to be changed to warnings when -a is passed.
     if (convertToWarnings.includes(resolvedKeyword) && additionalPropertiesAllowed) {
       updateWarningsCache({ warningsCache, resourceName, failedItemName, message, fileName, stats });
       return acc;
@@ -406,30 +435,13 @@ const generateErrorReport = ({
 };
 
 /**
- * TODO: move to common
- * @param {*} urn
- * @returns
- */
-const parseResoUrn = (urn = '') => {
-  const parts = urn?.split?.(':') || '';
-
-  if (parts.length < 6 || parts[0] !== 'urn' || parts[1] !== 'reso' || parts[2] !== 'metadata') {
-    return {
-      version: '',
-      resource: ''
-    };
-  }
-
-  return {
-    version: parts[3],
-    resource: parts.slice(5)[0]
-  };
-};
-
-/**
  * TODO
- * @param {*} param0
- * @returns
+ * @param {Object} obj
+ * @param {{}} obj.payload
+ * @param {string} obj.resource
+ * @param {string} obj.version
+ *
+ * @returns If the urn is a RESO urn this returns the resource and version parsed from that else returns the resource and version that was passed in.
  */
 const getResourceAndVersion = ({ payload, resource, version }) => {
   if (payload['@reso.context']) {
@@ -447,10 +459,18 @@ const getResourceAndVersion = ({ payload, resource, version }) => {
 };
 
 /**
- * TODO - convert to arrow function
- * @param {*} param0
+ *
+ * @param {Object} obj
+ * @param {{}} obj.errorCache
+ * @param {String} obj.resourceName
+ * @param {String} obj.failedItemName
+ * @param {String} obj.message
+ * @param {String} obj.fileName
+ * @param {{totalErrors: Number}} obj.stats
+ *
+ * Updates the cache with a new error message
  */
-function updateErrorCache({ errorCache, resourceName, failedItemName, message, fileName, stats }) {
+const updateErrorCache = ({ errorCache, resourceName, failedItemName, message, fileName, stats }) => {
   if (!errorCache[resourceName]) {
     errorCache[resourceName] = {};
   }
@@ -471,13 +491,21 @@ function updateErrorCache({ errorCache, resourceName, failedItemName, message, f
     errorCache[resourceName][failedItemName][`__${message}`][fileName].occurrences++;
   }
   stats.totalErrors++;
-}
+};
 
 /**
- * TODO - convert to arrow function
- * @param {*} param0
+ *
+ * @param {Object} obj
+ * @param {{}} obj.warningsCache
+ * @param {String} obj.resourceName
+ * @param {String} obj.failedItemName
+ * @param {String} obj.message
+ * @param {String} obj.fileName
+ * @param {{totalWarnings: Number}} obj.stats
+ *
+ * Updates the cache with a new warning message
  */
-function updateWarningsCache({ warningsCache, resourceName, failedItemName, message, fileName, stats }) {
+const updateWarningsCache = ({ warningsCache, resourceName, failedItemName, message, fileName, stats }) => {
   if (!warningsCache[resourceName]) {
     warningsCache[resourceName] = {};
   }
@@ -498,7 +526,7 @@ function updateWarningsCache({ warningsCache, resourceName, failedItemName, mess
     warningsCache[resourceName][failedItemName][`__${message}`][fileName].occurrences++;
   }
   stats.totalWarnings++;
-}
+};
 
 module.exports = {
   validatePayload,
